MSE can now save cards in separate files (needs manual config editing still). Trailing slashes are stripped from commandline arguments (because directory.mse-set/ should be accepted)
git-svn-id: svn://svn.code.sf.net/p/magicseteditor/code/trunk@1386 0fc631ac-6414-0410-93d0-97cfa31319b6

+++ src/gui/set/style_panel.cpp
@@ -70,7 +70,7 @@ void StylePanel::updateListSize() {
 	// we only need enough columns to show all items
 	int x_room = GetSize().x - editor->GetBestSize().x - 6;
 	size_t need_columns = (size_t)ceil(list->requiredWidth() / (double)x_room);
-	size_t column_count = max(1ul, min(fit_columns, need_columns));
+	size_t column_count = max((size_t)1, min(fit_columns, need_columns));
 	// change count
 	if (column_count != list->column_count) {
 		list->column_count = column_count;
